Created Event parser.
Now it is *only* the JSON converters...
